+++ baseband/vlbi_base/utils.py
@@ -2,6 +2,7 @@
 from __future__ import division, unicode_literals, print_function
 
 import numpy as np
+import six
 
 __all__ = ['bcd_decode', 'bcd_encode', 'CRC']
 
@@ -122,3 +123,17 @@ class CRC(object):
         for i in range(0, len(stream) - len(self)):
             stream[i:i+pol_bin.size] ^= (pol_bin & stream[i])
         return stream[-len(self):]
+
+
+if not six.PY2:  # ignoring python < 3.5; could add if you wish
+    from math import gcd
+else:
+    from fractions import gcd as fgcd
+
+    def gcd(a, b):
+        return abs(fgcd(a, b))
+
+
+def lcm(a, b):
+    """Calculate the least common multiple of a and b."""
+    return abs(a * b) // gcd(a, b)